+++ fooof/analysis/__init__.py
@@ -1,0 +1,4 @@
+"""Analysis sub-module for FOOOF."""
+
+from .error import compute_pointwise_error_fm, compute_pointwise_error_fg
+from .periodic import get_band_peak, get_band_peak_group, get_band_peak_fm, get_band_peak_fg
