+++ fooof/analysis/error.py
@@ -1,61 +1,14 @@
-"""Functions to check FOOOF models."""
+"""Functions to analyze and investigate FOOOF results - model fit error."""
 
 import numpy as np
 
 from fooof.sim.gen import gen_model
-from fooof.core.info import get_description
-from fooof.core.errors import NoModelError, NoDataError
 from fooof.utils import compute_pointwise_error
 from fooof.plts.spectra import plot_spectrum_error
+from fooof.core.errors import NoModelError, NoDataError
 
 ###################################################################################################
 ###################################################################################################
-
-def get_info(fooof_obj, aspect):
-    """Get a specified selection of information from a FOOOF derived object.
-
-    Parameters
-    ----------
-    fooof_obj : FOOOF or FOOOFGroup
-        Object to get attributes from.
-    aspect : {'settings', 'meta_data', 'results'}
-        Which set of attributes to compare the objects across.
-
-    Returns
-    -------
-    dict
-        The set of specified info from the FOOOF derived object.
-    """
-
-    return {key : getattr(fooof_obj, key) for key in get_description()[aspect]}
-
-
-def compare_info(fooof_lst, aspect):
-    """Compare a specified aspect of FOOOF objects across instances.
-
-    Parameters
-    ----------
-    fooof_lst : list of FOOOF or list of FOOOFGroup
-        Objects whose attributes are to be compared.
-    aspect : {'setting', 'meta_data'}
-        Which set of attributes to compare the objects across.
-
-    Returns
-    -------
-    consistent : bool
-        Whether the settings are consistent across the input list of objects.
-    """
-
-    # Check specified aspect of the objects are the same across instances
-    for f_obj_1, f_obj_2 in zip(fooof_lst[:-1], fooof_lst[1:]):
-        if getattr(f_obj_1, 'get_' + aspect)() != getattr(f_obj_2, 'get_' + aspect)():
-            consistent = False
-            break
-    else:
-        consistent = True
-
-    return consistent
-
 
 def compute_pointwise_error_fm(fm, plot_errors=True, return_errors=False, **plt_kwargs):
     """Calculate the frequency by frequency error of a model fit from a FOOOF object.
