+++ fooof/data/__init__.py
@@ -1,0 +1,4 @@
+"""Data sub-module for FOOOF."""
+
+from .bands import Bands
+from .data import FOOOFSettings, FOOOFMetaData, FOOOFResults, SimParams
