+++ fooof/data/bands.py
