+++ fooof/objs/__init__.py
@@ -1,0 +1,5 @@
+"""Objects sub-module, for FOOOF objects and functions that operate on FOOOF objects."""
+
+from .fit import FOOOF
+from .group import FOOOFGroup
+from .utils import average_fg, combine_fooofs, fit_fooof_group_3d
