+++ fooof/tests/test_settings.py
@@ -1,10 +1,12 @@
 """Settings for FOOOF tests."""
 
+import os
 import pkg_resources as pkg
 
 ###################################################################################################
 ###################################################################################################
 
 # Path Settings
-TEST_FILE_PATH = pkg.resource_filename(__name__, 'test_files')
-TEST_REPORTS_PATH = pkg.resource_filename(__name__, 'test_reports')
+BASE_TEST_FILE_PATH = pkg.resource_filename(__name__, 'test_files')
+TEST_FILE_PATH = os.path.join(BASE_TEST_FILE_PATH, 'data')
+TEST_REPORTS_PATH = os.path.join(BASE_TEST_FILE_PATH, 'reports')
