+++ fooof/tests/test_utils.py
@@ -2,10 +2,10 @@
 
 from functools import wraps
 
-from fooof import FOOOF, FOOOFGroup
-from fooof.bands import Bands
-from fooof.sim.gen import gen_power_spectrum, gen_group_power_spectra
+from fooof.data.bands import Bands
+from fooof.objs import FOOOF, FOOOFGroup
 from fooof.sim.params import param_sampler
+from fooof.sim.gen import gen_power_spectrum, gen_group_power_spectra
 from fooof.core.modutils import safe_import
 
 plt = safe_import('.pyplot', 'matplotlib')
