+++ fooof/utils/__init__.py
@@ -1,0 +1,4 @@
+"""Utilities sub-module for FOOOF."""
+
+from .debug import sys_info
+from .data import trim_spectrum, compute_pointwise_error
